response to legobeat review
